requirements.txt
mysql-connector-python==9.1.0
python-dotenv==1.0.1
setuptools==75.6.0
tabulate==0.9.0
wheel==0.45.1
